add getting ahead of myself again.
I started pasting in computed values pulled from devtools on the google page, lost track of what was happening, ended up with way too much duplicated styling, and basically up a creek without a paddle. I will go back and do it the right way, styling everything on my own, not necessarily in the same way that the google page is styled, but to achieve the same outcome.